+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Upgrading to 3.6.x versions requires that you upgrade to latest 3.5.x version fi
 
 - Remove last PEAR dependency: `Math_Stats` (@glensc, #509)
 - Add boot and shutdown events (@glensc, #511)
+- Regression: Fix notification `role_id` type notice (@glensc, #513)
 
 [3.6.4]: https://github.com/eventum/eventum/compare/v3.6.3...master